requirements.txt
Django==2.1.2
gunicorn
whitenoise
#pylibmc==1.5.2
#django-heroku-memcacheify
python-docx
python-dateutil
rfc3339
